Todoist: Create Toggl project when creating Todoist project
When adding a project in the Todoist sidebar, hold Alt when pressing Enter
to also create a Toggl project with that name. The nearest parent
Todoist project that matches a Toggl client is used as the client for the
new Toggl project.

If a Toggl project with that name already exists for that client, no new
project is created. If no client matches one of the Todoist parent projects,
Toggl projects for all clients are searched to find a match.