Todoist: Specify workspace and/or client when creating project
This update gives the ability to specify the workspace and client when
creating a Toggl project via the Alt+Enter submission of a project name
in the Todoist sidebar.

To specify the client, create a Todoist project with the client name
and nest as a child the project representing the Toggl project.

To specify the workspace, create a Todoist project with the workspace
name and nest as a child the project representing the Toggl client or
project.

For example, the projects in the sidebar should be arranged this way:
 * Workspace
 +--* Client (optional)
    +--* [Project  ]

Where 'Project' is the input box where you specify the name of the Toggl
project, and where you press Alt+Enter to create the Toggl project.

Implementation notes:
 * In the Todoist project hierarchy, Client need not be a direct child of
Workspace, nor Project of Client. Intermediate parents are allowed.
 * Note also that we search the ancestry line from the top down for
Workspaces but from the bottom up for the Client.
 * If a workspace is found in the Todoist project hierarchy, the
subsequent search for clients will only match clients that are in that
workspace.